# sha2: add `read_volatile` workaround for round constants