--- a/sha2/src/sha256/soft.rs
+++ b/sha2/src/sha256/soft.rs
@@ -1,5 +1,5 @@
 #![allow(clippy::many_single_char_names)]
-use crate::consts::BLOCK_LEN;
+use crate::consts::{BLOCK_LEN, K32};
 
 #[inline(always)]
 fn shr(v: [u32; 4], o: u32) -> [u32; 4] {
@@ -28,6 +28,33 @@ fn add(a: [u32; 4], b: [u32; 4]) -> [u32; 4] {
         a[1].wrapping_add(b[1]),
         a[2].wrapping_add(b[2]),
         a[3].wrapping_add(b[3]),
+    ]
+}
+
+#[inline(always)]
+fn add_round_const(a: [u32; 4], i: usize) -> [u32; 4] {
+    #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
+    fn k(i: usize, j: usize) -> u32 {
+        // `read_volatile` forces compiler to read round constants from the static
+        // instead of inlining them, which improves codegen and performance on some platforms.
+        // On x86 targets 32-bit constants can be encoded using immediate argument on the `add`
+        // instruction, so it's more efficient to inline them.
+        cfg_if::cfg_if! {
+            if #[cfg(any(target_arch = "x86", target_arch = "x86_64"))] {
+                use core::ptr::read as r;
+            } else {
+                use core::ptr::read_volatile as r;
+            }
+        }
+
+        unsafe { r(K32.as_ptr().add(4 * i + j)) }
+    }
+
+    [
+        a[0].wrapping_add(k(i, 3)),
+        a[1].wrapping_add(k(i, 2)),
+        a[2].wrapping_add(k(i, 1)),
+        a[3].wrapping_add(k(i, 0)),
     ]
 }
 
@@ -142,7 +169,7 @@ fn schedule(v0: [u32; 4], v1: [u32; 4], v2: [u32; 4], v3: [u32; 4]) -> [u32; 4] 
 
 macro_rules! rounds4 {
     ($abef:ident, $cdgh:ident, $rest:expr, $i:expr) => {{
-        let t1 = add($rest, crate::consts::K32X4[$i]);
+        let t1 = add_round_const($rest, $i);
         $cdgh = sha256_digest_round_x2($cdgh, $abef, t1);
         let t2 = sha256swap(t1);
         $abef = sha256_digest_round_x2($abef, $cdgh, t2);