Prevents compiler from inlining round constants or spilling them to stack, which can slightly improve performance.

For example, on my x86-64 laptop:
```
// Before:
test sha512_10    ... bench:          23 ns/iter (+/- 0) = 434 MB/s
test sha512_100   ... bench:         208 ns/iter (+/- 1) = 480 MB/s
test sha512_1000  ... bench:       2,015 ns/iter (+/- 7) = 496 MB/s
test sha512_10000 ... bench:      20,047 ns/iter (+/- 862) = 498 MB/s

// After:
test sha512_10    ... bench:          22 ns/iter (+/- 0) = 454 MB/s
test sha512_100   ... bench:         201 ns/iter (+/- 0) = 497 MB/s
test sha512_1000  ... bench:       1,941 ns/iter (+/- 3) = 515 MB/s
test sha512_10000 ... bench:      19,298 ns/iter (+/- 64) = 518 MB/s
``` 

Closes #328