@newpavlov 2024-01-12
@piegamesde
Can you check that this PR indeed helps with performance issues on RISC-V?