@tarcieri 2024-01-12
> or spilling them to stack

Interesting. Wonder if we could use that approach in the `aes` crate as well?