@newpavlov 2024-01-12
>Wonder if we could use that approach in the aes crate as well?

It will be a bit harder to do. If there is enough registers, we want keys to stay in them, while `read_volatile` would always read them. In the `sha2` case we know for certain that there is not enough registers, even on register rich targets like RISC-V, so reading round constants on every block processing is fine.